send media and logo data to backend
