+++ src/app/api/uploadthing/core.ts
@@ -2,7 +2,7 @@ import { createUploadthing, type FileRouter } from "uploadthing/next";
 import { UploadThingError } from "uploadthing/server";
 import { auth } from "@clerk/nextjs/server";
 import { db } from "~/server/db";
-import { z } from "zod";
+import { number, z } from "zod";
 import { business, media, user } from "~/server/db/schema";
 
 const f = createUploadthing();
@@ -16,7 +16,6 @@ export const ourFileRouter = {
     .middleware(async ({ req, input }) => {
       // This code runs on your server before upload
       const user = auth();
-
       // If you throw, the user will not be able to upload
       // eslint-disable-next-line @typescript-eslint/only-throw-error
       if (!user) throw new UploadThingError("Unauthorized");
@@ -36,39 +35,37 @@ export const ourFileRouter = {
         name: file.name,
         businessID: 3,
       });
-
-      // businessID: serial("businessID").references(() => business.businessID),
-      //     userID: varchar("userID", {length: 256}).references(() => user.userID),
-      //     url: varchar("url", {length: 1024}).notNull(),
-      //     name: varchar("name", {length: 256}).notNull(),
-      // !!! Whatever is returned here is sent to the clientside `onClientUploadComplete` callback
+      console.log("insert complete");
       return { uploadedBy: metadata.userId };
     }),
-  // // Define as many FileRoutes as you like, each with a unique routeSlug
-  // videoUploader: f({ "video/mp4": { maxFileSize: "16MB" } })
-  //     // Set permissions and file types for this FileRoute
-  //     .middleware(async ({ req }) => {
-  //         // This code runs on your server before upload
-  //         const user = auth();
-  //         // If you throw, the user will not be able to upload
-  //         if (!user) throw new UploadThingError("Unauthorized");
-  //
-  //         // Whatever is returned here is accessible in onUploadComplete as `metadata`
-  //         return { userId: user.userId };
-  //     })
-  //     .onUploadComplete(async ({ metadata, file }) => {
-  //         // This code RUNS ON YOUR SERVER after upload
-  //         console.log("Upload complete for userId:", metadata.userId);
-  //
-  //         console.log("file url", file.url);
-  //         await db.insert(image).values({
-  //             name: file.name,
-  //             url: file.url,
-  //             userId: metadata.userId ?? "",
-  //         });
-  //         // !!! Whatever is returned here is sent to the clientside `onClientUploadComplete` callback
-  //         return { uploadedBy: metadata.userId };
-  //     }),
+
+  logoUploader: f(["image/png", "image/jpeg"])
+    .input(z.string())
+    .middleware(async ({ req, input }) => {
+      // This code runs on your server before upload
+      const user = auth();
+      // If you throw, the user will not be able to upload
+      // eslint-disable-next-line @typescript-eslint/only-throw-error
+      if (!user) throw new UploadThingError("Unauthorized");
+      // Whatever is returned here is accessible in onUploadComplete as `metadata`
+      return { userId: user.userId, input };
+    })
+    .onUploadComplete(async ({ metadata, file }) => {
+      // This code RUNS ON YOUR SERVER after upload
+      console.log("Upload complete for userId:", metadata.userId);
+      console.log(metadata.input, "test test");
+      console.log("file url", file.url);
+      console.log("writing to db");
+
+      await db.insert(media).values({
+        userID: metadata.userId ?? "",
+        url: file.url,
+        name: file.name,
+        businessID: 3,
+      });
+      console.log("insert complete");
+      return { uploadedBy: metadata.userId };
+    }),
 } satisfies FileRouter;
 
 export type OurFileRouter = typeof ourFileRouter;
