+++ src/components/uploadbuttton.tsx
@@ -23,3 +23,23 @@ export function UploadButtonCom() {
     />
   );
 }
+
+export function LogoUploadButton() {
+  const router = useRouter();
+  return (
+    <UploadButton
+      input=""
+      endpoint="logoUploader"
+      onClientUploadComplete={(res) => {
+        // Do something with the response
+        console.log("Files: ", res);
+        alert("Upload Completed");
+        router.refresh();
+      }}
+      onUploadError={(error: Error) => {
+        // Do something with the error.
+        alert(`ERROR! ${error.message}`);
+      }}
+    />
+  );
+}
